Updated useQuery hook, removed CurrentUser Type and moved files
