+++ client/src/attributes.ts
@@ -1,10 +1,3 @@
-export const CURRENT_USER_ATTRIBUTES = [
-  "id",
-  "name",
-  "profilePicture",
-  "chatRooms",
-];
-
 export const USER_ATTRIBUTES = ["id", "name", "profilePicture"];
 
 export const CHATROOM_MIN_ATTRIBUTES = ["id", "type", "name"];
