+++ client/src/components/NewChatRoomForm/useNewChatRoom.ts
@@ -1,4 +1,4 @@
-import { useInput } from "pages/Login/Input/useInput";
+import { useInput } from "hooks/useInput";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { SelectType } from "./Select/Select";
 import { useSelect } from "./Select/useSelect";
