+++ client/src/pages/ChatRoom/ChatRoom.tsx
@@ -1,11 +1,11 @@
 import React, { useState, useCallback, useContext } from "react";
 import "./chat-room.css";
-import { ajaxClient } from "ajaxClient";
+
 import { CurrentUserContext } from "contexts/CurrentUserContext";
 import { CHATROOM_MIN_ATTRIBUTES } from "attributes";
 import { ChatRoomMin } from "types/ChatRoom";
 import { useQuery } from "hooks/useQuery";
-import Modal from "components/Modal/Modal";
+import { Modal } from "components/Modal/Modal";
 import NewChatRoomForm from "components/NewChatRoomForm/NewChatRoomForm";
 import { RoomList } from "./RoomList/RoomList";
 import { ChatArea } from "./ChatArea/ChatArea";
@@ -25,20 +25,15 @@ export const ChatRoom = () => {
     setShowNewChatRoomDialog(false);
   }, []);
 
-  const fetchChatRoomsMin = useCallback(
-    () =>
-      ajaxClient.get(
-        `/users/${currentUser?.id}/chats?fields=${CHATROOM_MIN_ATTRIBUTES.join(
-          ","
-        )}`
-      ),
-    [currentUser?.id]
+  const chatRoomListMinQuery = useQuery<ChatRoomMin[]>(
+    `/users/${currentUser?.id}/chats?fields=${CHATROOM_MIN_ATTRIBUTES.join(
+      ","
+    )}`,
+    {
+      enabled: true,
+      refetchInterval: 2,
+    }
   );
-
-  const chatRoomListMinQuery = useQuery<ChatRoomMin[]>(fetchChatRoomsMin, {
-    enabled: true,
-    refetchInterval: 2,
-  });
 
   const chatRoomListMin = chatRoomListMinQuery.data;
 
@@ -46,11 +41,6 @@ export const ChatRoom = () => {
     setChatRoomID(id);
   }, []);
 
-  // if (
-  //   chatRoomListMinQuery.status === "loading" ||
-  //   chatRoomListMinQuery.status === "idle"
-  // )
-  //   return <p>Loading....</p>;
   if (chatRoomListMinQuery.status === "error")
     return (
       <p>{chatRoomListMinQuery.error?.message || "Some Error Occured...."}</p>
