+++ client/src/pages/Login/Login.tsx
@@ -3,9 +3,9 @@ import React, { FormEvent, useCallback, useEffect } from "react";
 import { GrLogin } from "react-icons/gr";
 import { AiOutlineMail } from "react-icons/ai";
 import { RiLockPasswordFill } from "react-icons/ri";
-import { useInput } from "./Input/useInput";
+import { useInput } from "../../hooks/useInput";
 import "./login.css";
-import { Input } from "./Input/Input";
+import { Input } from "../../components/Input/Input";
 import { ajaxClient } from "ajaxClient";
 import { useMutation } from "hooks/useMutation";
 import { Button } from "components/Button/Button";
